# Conflicts:
#	Backend/AnyTrack.Backend/Service/SprintService.cs

+++ Backend/AnyTrack.Backend/Service/SprintService.cs
@@ -3,6 +3,7 @@ using System.CodeDom;
 using System.Collections.Generic;
 using System.Data.Entity.Infrastructure;
 using System.Linq;
+using System.Net.Mail;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -289,7 +290,7 @@ namespace AnyTrack.Backend.Service
             foreach (var dataTask in tasks)
             {
                 var remainingTaskHours =
-                    unitOfWork.TaskHourEstimateRepository.Items.Where(t => t.Id == dataTask.Id).ToList();
+                    unitOfWork.TaskHourEstimateRepository.Items.Where(t => t.Task.Id == dataTask.Id).ToList();
 
                 List<ServiceTaskHourEstimate> serviceRemainingTaskHours = new List<ServiceTaskHourEstimate>();
                 foreach (var dataRemainingTaskHours in remainingTaskHours)
@@ -297,7 +298,8 @@ namespace AnyTrack.Backend.Service
                     ServiceTaskHourEstimate serviceTaskHourEstimate = new ServiceTaskHourEstimate()
                     {
                         Estimate = dataRemainingTaskHours.Estimate,
-                        TaskId = dataTask.Id    
+                        TaskId = dataTask.Id,
+                        Created = dataRemainingTaskHours.Created
                     };
                     serviceRemainingTaskHours.Add(serviceTaskHourEstimate);
                 }
@@ -487,9 +489,10 @@ namespace AnyTrack.Backend.Service
         /// <summary>
         /// All stories in the backlog right now
         /// </summary>
+        /// <param name="projectId">The project id</param>
         /// <param name="sprintId">The sprint id</param>
         /// <param name="sprintStories">The Sprint stories</param>
-        public void ManageSprintBacklog(Guid sprintId, List<ServiceSprintStory> sprintStories)
+        public void ManageSprintBacklog(Guid projectId, Guid sprintId, List<ServiceSprintStory> sprintStories)
         {
             if (sprintId == null)
             {
@@ -545,19 +548,56 @@ namespace AnyTrack.Backend.Service
                 ////Then we need to remove this from the repo
                 if (!newSprintIds.Contains(story.Story.Id))
                 {
-                    var productBacklogStory = unitOfWork.StoryRepository.Items.Single(s => s.Id == story.Story.Id);
-                    productBacklogStory.InSprint = false;
                     removeStories.Add(story);
                 }
             }
 
             foreach (var story in removeStories)
             {
+                var project = unitOfWork.ProjectRepository.Items.SingleOrDefault(p => p.Id == projectId);
+
+                if (project == null)
+                {
+                    throw new NullReferenceException("project");
+                }
+
+                var dataStory = project.Stories.SingleOrDefault(s => s.Id == story.Story.Id);
+                if (dataStory == null)
+                {
+                    throw new NullReferenceException("dataStory");
+                }
+
+                dataStory.InSprint = false;
+                unitOfWork.Commit();
+                                
                 var sprintStory = unitOfWork.SprintStoryRepository.Items.Single(s => s.Story.Id == story.Story.Id);
                 unitOfWork.SprintStoryRepository.Delete(sprintStory);
             }
 
             unitOfWork.Commit();
+        }
+
+        /// <summary>
+        /// Sends an email request
+        /// </summary>
+        /// <param name="senderEmailAddress">The sender email address</param>
+        /// <param name="recipientEmailAddress">The recipient email message</param>
+        /// <param name="emailMessage">The email address</param>
+        /// <param name="emailAttachment">The email attachment</param>
+        public void SendEmailRequest(string senderEmailAddress, string recipientEmailAddress, string emailMessage, Attachment emailAttachment)
+        {
+            MailMessage mail = new MailMessage();
+
+            mail.From = new MailAddress(senderEmailAddress);
+            mail.To.Add(recipientEmailAddress);
+
+            mail.Subject = "Agile";
+            mail.Body = emailMessage;
+
+            mail.Attachments.Add(emailAttachment);
+
+            SmtpClient smtp = new SmtpClient("127.0.0.1");
+            smtp.Send(mail);
         }
 
         #endregion
